requirements.txt
pandas 
flask
numpy 
matplotlib
seaborn 
scikit-learn
scipy
mysql-connector-python
python-dotenv
pymysql
xlrd
openpyxl
ipykernel
xgboost
catboost
dill
mlflow
dagshub

#-e .